feat(backend): analyze 加 save_image + _save_screenshot 提位到 backend
